refactor: move Explore component to /projectsV2
