+++ src/features/projectsV2/ProjectsMap/Explore/CustomSlider.tsx
@@ -1,5 +1,5 @@
 import { Slider, styled } from '@mui/material';
-import theme from '../../theme/themeProperties';
+import theme from '../../../../theme/themeProperties';
 const { light, nonDonatableProjectBackgroundColor } = theme;
 export const SmallSlider = styled(Slider)(() => ({
   '&.MuiSlider-root': {
