+++ src/features/projectsV2/ProjectsMap/Explore/Explore.module.scss
@@ -1,4 +1,4 @@
-@import '../../theme/theme';
+@import '../../../../theme/theme';
 .exploreMainContainer {
   max-width: 182px;
   background-color: $backgroundColor;
